Move _make_attachment to trait WP_Test_RunsFileUploadTests.
